Reproduce

+++ src/infrastructure/transformers.py
@@ -1,4 +1,5 @@
 from transformers import AutoTokenizer, PreTrainedTokenizer, AutoModelForSeq2SeqLM, PreTrainedModel
+from torch import float16
 
 
 def load_tokenizer(tokenizer_id: str) -> PreTrainedTokenizer:
@@ -6,4 +7,4 @@ def load_tokenizer(tokenizer_id: str) -> PreTrainedTokenizer:
 
 
 def load_base_model(base_model_id: str) -> PreTrainedModel:
-    return AutoModelForSeq2SeqLM.from_pretrained(base_model_id, load_in_8bit=True, device_map="auto")
+    return AutoModelForSeq2SeqLM.from_pretrained(base_model_id, load_in_8bit=True, device_map="auto", torch_dtype=float16)
